+++ tests/Kiota.Builder.Tests/CodeDOM/CodeClassTests.cs
@@ -50,7 +50,7 @@ public class CodeClassTests
         {
             Name = "class1"
         }).First();
-        codeClass.Indexer = new CodeIndexer
+        codeClass.AddIndexer(new CodeIndexer
         {
             Name = "idx",
             SerializationName = "idx_smth",
@@ -67,13 +67,13 @@ public class CodeClassTests
                 IsNullable = true,
             },
             IndexParameterName = "idxSmth",
-        };
+        });
         Assert.Single(codeClass.GetChildElements(true).OfType<CodeIndexer>());
         Assert.Throws<ArgumentNullException>(() =>
         {
-            codeClass.Indexer = null;
+            codeClass.AddIndexer(null);
         });
-        codeClass.Indexer = new CodeIndexer
+        codeClass.AddIndexer(new CodeIndexer
         {
             Name = "idx2",
             SerializationName = "idx-2",
@@ -90,9 +90,9 @@ public class CodeClassTests
                 IsNullable = true,
             },
             IndexParameterName = "idx2",
-        };
+        });
         Assert.Empty(codeClass.GetChildElements(true).OfType<CodeIndexer>());
-        var methods = codeClass.GetChildElements(true).OfType<CodeMethod>().Where(x => x.IsOfKind(CodeMethodKind.IndexerBackwardCompatibility)).ToArray();
+        var methods = codeClass.GetChildElements(true).OfType<CodeMethod>().Where(static x => x.IsOfKind(CodeMethodKind.IndexerBackwardCompatibility)).ToArray();
         Assert.Equal(2, methods.Length);
         Assert.Equal("WithIdxSmth", methods.FirstOrDefault(static x => x.OriginalIndexer.Name.Equals("idx")).Name);
         Assert.Equal("WithIdx2", methods.FirstOrDefault(static x => x.OriginalIndexer.Name.Equals("idx2")).Name);